Refactor to DDD structure
